pyproject.toml
[project]
name = "obelisk-recontext-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Jeremy Wortz",email = "jwortz@google.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "python-dotenv",
    "flask",
    "google-auth",
    "google-adk (>=1.5.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "obelisk-recontext-agent"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
